remove repeated codes
