+++ log.go
@@ -1,13 +1,9 @@
-// +build !windows
-
 package main
 
 import (
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"log/syslog"
 	"os"
 	"path"
 	"strconv"
@@ -39,12 +35,6 @@ type FileLogger struct {
 	file            *os.File
 	logEventEmitter LogEventEmitter
 	locker          sync.Locker
-}
-
-type SysLogger struct {
-	NullLogger
-	logWriter       *syslog.Writer
-	logEventEmitter LogEventEmitter
 }
 
 type NullLogger struct {
@@ -311,31 +301,6 @@ func (l *FileLogger) Close() error {
 	return nil
 }
 
-func NewSysLogger(name string, logEventEmitter LogEventEmitter) *SysLogger {
-	writer, err := syslog.New(syslog.LOG_DEBUG, name)
-	logger := &SysLogger{logEventEmitter: logEventEmitter}
-	if err == nil {
-		logger.logWriter = writer
-	}
-	return logger
-}
-
-func (sl *SysLogger) Write(b []byte) (int, error) {
-	sl.logEventEmitter.emitLogEvent(string(b))
-	if sl.logWriter != nil {
-		return sl.logWriter.Write(b)
-	} else {
-		return 0, errors.New("not connect to syslog server")
-	}
-}
-
-func (sl *SysLogger) Close() error {
-	if sl.logWriter != nil {
-		return sl.logWriter.Close()
-	} else {
-		return errors.New("not connect to syslog server")
-	}
-}
 func NewNullLogger(logEventEmitter LogEventEmitter) *NullLogger {
 	return &NullLogger{logEventEmitter: logEventEmitter}
 }
