+++ logger_windows.go
@@ -1,0 +1,14 @@
+// +build windows
+
+package main
+
+type SysLogger struct {
+	NullLogger
+	logEventEmitter LogEventEmitter
+}
+
+func NewSysLogger(name string, logEventEmitter LogEventEmitter) *SysLogger {
+	return &SysLogger{
+		logEventEmitter: logEventEmitter,
+	}
+}
